some pr feedback
