+++ service/src/main/java/bio/terra/workspace/app/configuration/external/AzureConfiguration.java
@@ -12,6 +12,8 @@ public class AzureConfiguration {
   private String managedAppClientId;
   private String managedAppClientSecret;
   private String managedAppTenantId;
+  private Long sasTokenStartTimeMinutesOffset;
+  private Long sasTokenExpiryTimeMinutesOffset;
 
   public String getManagedAppClientId() {
     return managedAppClientId;
@@ -35,5 +37,17 @@ public class AzureConfiguration {
 
   public void setManagedAppTenantId(String managedAppTenantId) {
     this.managedAppTenantId = managedAppTenantId;
+  }
+
+  public Long getSasTokenStartTimeMinutesOffset() { return sasTokenStartTimeMinutesOffset; }
+
+  public void setSasTokenStartTimeMinutesOffset(Long sasTokenStartTimeMinutesOffset) {
+    this.sasTokenStartTimeMinutesOffset = sasTokenStartTimeMinutesOffset;
+  }
+
+  public Long getSasTokenExpiryTimeMinutesOffset() { return sasTokenExpiryTimeMinutesOffset; }
+
+  public void setSasTokenExpiryTimeMinutesOffset(Long sasTokenExpiryTimeMinutesOffset) {
+    this.sasTokenExpiryTimeMinutesOffset = sasTokenExpiryTimeMinutesOffset;
   }
 }
