+++ service/src/main/java/bio/terra/workspace/app/controller/ControlledAzureResourceApiController.java
@@ -1,10 +1,34 @@
 package bio.terra.workspace.app.controller;
 
 import bio.terra.common.exception.ApiException;
+import bio.terra.workspace.app.configuration.external.AzureConfiguration;
 import bio.terra.workspace.app.configuration.external.FeatureConfiguration;
 import bio.terra.workspace.common.utils.AzureVmUtils;
 import bio.terra.workspace.generated.controller.ControlledAzureResourceApi;
-import bio.terra.workspace.generated.model.*;
+import bio.terra.workspace.generated.model.ApiAzureDiskResource;
+import bio.terra.workspace.generated.model.ApiAzureIpResource;
+import bio.terra.workspace.generated.model.ApiAzureNetworkResource;
+import bio.terra.workspace.generated.model.ApiAzureRelayNamespaceResource;
+import bio.terra.workspace.generated.model.ApiAzureVmResource;
+import bio.terra.workspace.generated.model.ApiCreateControlledAzureDiskRequestBody;
+import bio.terra.workspace.generated.model.ApiCreateControlledAzureIpRequestBody;
+import bio.terra.workspace.generated.model.ApiCreateControlledAzureNetworkRequestBody;
+import bio.terra.workspace.generated.model.ApiCreateControlledAzureRelayNamespaceRequestBody;
+import bio.terra.workspace.generated.model.ApiCreateControlledAzureRelayNamespaceResult;
+import bio.terra.workspace.generated.model.ApiCreateControlledAzureStorageContainerRequestBody;
+import bio.terra.workspace.generated.model.ApiCreateControlledAzureStorageRequestBody;
+import bio.terra.workspace.generated.model.ApiCreateControlledAzureVmRequestBody;
+import bio.terra.workspace.generated.model.ApiCreatedAzureStorageContainerSasToken;
+import bio.terra.workspace.generated.model.ApiCreatedControlledAzureDisk;
+import bio.terra.workspace.generated.model.ApiCreatedControlledAzureIp;
+import bio.terra.workspace.generated.model.ApiCreatedControlledAzureNetwork;
+import bio.terra.workspace.generated.model.ApiCreatedControlledAzureStorage;
+import bio.terra.workspace.generated.model.ApiCreatedControlledAzureStorageContainer;
+import bio.terra.workspace.generated.model.ApiCreatedControlledAzureVmResult;
+import bio.terra.workspace.generated.model.ApiDeleteControlledAzureResourceRequest;
+import bio.terra.workspace.generated.model.ApiDeleteControlledAzureResourceResult;
+import bio.terra.workspace.generated.model.ApiJobControl;
+import bio.terra.workspace.generated.model.ApiJobReport;
 import bio.terra.workspace.service.iam.AuthenticatedUserRequest;
 import bio.terra.workspace.service.iam.AuthenticatedUserRequestFactory;
 import bio.terra.workspace.service.iam.SamRethrow;
@@ -12,6 +36,7 @@ import bio.terra.workspace.service.iam.SamService;
 import bio.terra.workspace.service.job.JobService;
 import bio.terra.workspace.service.resource.ResourceValidationUtils;
 import bio.terra.workspace.service.resource.controlled.ControlledResourceService;
+import bio.terra.workspace.service.resource.controlled.cloud.azure.AzureControlledStorageResourceService;
 import bio.terra.workspace.service.resource.controlled.cloud.azure.disk.ControlledAzureDiskResource;
 import bio.terra.workspace.service.resource.controlled.cloud.azure.ip.ControlledAzureIpResource;
 import bio.terra.workspace.service.resource.controlled.cloud.azure.network.ControlledAzureNetworkResource;
@@ -39,21 +64,27 @@ public class ControlledAzureResourceApiController extends ControlledResourceCont
   private final Logger logger = LoggerFactory.getLogger(ControlledGcpResourceApiController.class);
 
   private final ControlledResourceService controlledResourceService;
+  private final AzureControlledStorageResourceService azureControlledStorageResourceService;
   private final JobService jobService;
   private final FeatureConfiguration features;
+  private final AzureConfiguration azureConfiguration;
 
   @Autowired
   public ControlledAzureResourceApiController(
       AuthenticatedUserRequestFactory authenticatedUserRequestFactory,
       ControlledResourceService controlledResourceService,
+      AzureControlledStorageResourceService azureControlledStorageResourceService,
       SamService samService,
       JobService jobService,
       HttpServletRequest request,
-      FeatureConfiguration features) {
+      FeatureConfiguration features,
+      AzureConfiguration azureConfiguration) {
     super(authenticatedUserRequestFactory, request, controlledResourceService, samService);
     this.controlledResourceService = controlledResourceService;
+    this.azureControlledStorageResourceService = azureControlledStorageResourceService;
     this.jobService = jobService;
     this.features = features;
+    this.azureConfiguration = azureConfiguration;
   }
 
   @Override
@@ -186,10 +217,12 @@ public class ControlledAzureResourceApiController extends ControlledResourceCont
         storageContainerUuid.toString(),
         workspaceUuid.toString());
 
-    OffsetDateTime startTime = OffsetDateTime.now().minusMinutes(15);
-    OffsetDateTime expiryTime = OffsetDateTime.now().plusHours(1);
+    OffsetDateTime startTime =
+        OffsetDateTime.now().minusMinutes(azureConfiguration.getSasTokenStartTimeMinutesOffset());
+    OffsetDateTime expiryTime =
+        OffsetDateTime.now().plusMinutes(azureConfiguration.getSasTokenExpiryTimeMinutesOffset());
     String sas =
-        controlledResourceService.createAzureStorageContainerSasToken(
+        azureControlledStorageResourceService.createAzureStorageContainerSasToken(
             workspaceUuid, storageContainerUuid, startTime, expiryTime, userRequest);
 
     logger.info(
